requirements.txt
pandas==2.2.3
pyarrow==17.0.0
beautifulsoup4==4.12.3
lxml==5.3.0
ftfy==6.3.1
regex==2024.11.6
python-dateutil==2.9.0.post0
tqdm==4.67.1
typesense==0.21.0
python-dotenv==1.0.1
indic-transliteration==2.3.59